chore(model): RPC conversion code isolation (machine)
