+++ crates/api-model/src/machine/machine_search_config.rs
@@ -16,7 +16,6 @@
  */
 use carbide_uuid::instance_type::InstanceTypeId;
 use carbide_uuid::rack::RackId;
-use rpc::errors::RpcDataConversionError;
 
 /// MachineSearchConfig: Search parameters
 #[derive(Default, Debug, Clone)]
@@ -49,32 +48,4 @@ pub struct MachineSearchConfig {
     pub rack_id: Option<RackId>,
     /// Filter by the top-level controller state tag (e.g. "created", "ready")
     pub controller_state: Option<String>,
-}
-
-impl TryFrom<rpc::forge::MachineSearchConfig> for MachineSearchConfig {
-    type Error = RpcDataConversionError;
-
-    fn try_from(value: rpc::forge::MachineSearchConfig) -> Result<Self, Self::Error> {
-        Ok(MachineSearchConfig {
-            include_dpus: value.include_dpus,
-            include_history: value.include_history,
-            include_predicted_host: value.include_predicted_host,
-            only_maintenance: value.only_maintenance,
-            only_quarantine: value.only_quarantine,
-            exclude_hosts: value.exclude_hosts,
-            instance_type_id: value
-                .instance_type_id
-                .map(|t| {
-                    t.parse::<InstanceTypeId>()
-                        .map_err(|_| RpcDataConversionError::InvalidInstanceTypeId(t.clone()))
-                })
-                .transpose()?,
-            for_update: false, // This isn't exposed to API callers
-            mnnvl_only: value.mnnvl_only,
-            only_with_power_state: value.only_with_power_state,
-            only_with_health_alert: value.only_with_health_alert,
-            rack_id: value.rack_id,
-            controller_state: None,
-        })
-    }
 }
