+++ crates/api-model/src/machine/nvlink.rs
@@ -48,31 +48,6 @@ pub struct MachineNvLinkGpuStatusObservation {
     pub guid: u64,
 }
 
-impl From<MachineNvLinkStatusObservation> for rpc::forge::MachineNvLinkStatusObservation {
-    fn from(value: MachineNvLinkStatusObservation) -> Self {
-        rpc::forge::MachineNvLinkStatusObservation {
-            gpu_status: value
-                .nvlink_gpus
-                .into_iter()
-                .map(rpc::forge::MachineNvLinkGpuStatusObservation::from)
-                .collect(),
-        }
-    }
-}
-
-impl From<MachineNvLinkGpuStatusObservation> for rpc::forge::MachineNvLinkGpuStatusObservation {
-    fn from(value: MachineNvLinkGpuStatusObservation) -> Self {
-        rpc::forge::MachineNvLinkGpuStatusObservation {
-            gpu_id: value.gpu_id,
-            partition_id: value.partition_id,
-            logical_partition_id: value.logical_partition_id,
-            device_instance: value.device_instance,
-            domain_id: Some(value.domain_id),
-            guid: value.guid,
-        }
-    }
-}
-
 #[derive(Debug, Clone)]
 pub struct NvLinkConfigNotSyncedReason(pub String);
