+++ crates/api/src/handlers/bmc_endpoint_explorer.rs
@@ -24,9 +24,9 @@ use db::machine_interface::find_by_ip;
 use libredfish::RoleId;
 use mac_address::MacAddress;
 use model::expected_entity::ExpectedEntity;
-use model::machine::machine_id::try_parse_machine_id;
 use model::machine::machine_search_config::MachineSearchConfig;
 use model::machine::{LoadSnapshotOptions, MachineInterfaceSnapshot};
+use model::rpc_conv::machine::machine_id::try_parse_machine_id;
 use model::site_explorer::PreingestionState;
 use sqlx::PgConnection;
 use tokio::net::lookup_host;
